Refactor RegistraitonPage
